+++ src/sub_commands/broadcast_events.rs
@@ -5,9 +5,9 @@ use crate::utils::{create_client, handle_keys};
 
 #[derive(Args)]
 pub struct BroadcastEventsSubCommand {
-    /// Text note content
+    /// Input file path, should contain an array of JSON events
     #[arg(short, long)]
-    input: String,
+    file_path: String,
     // Print keys as hex
     #[arg(long, default_value = "false")]
     hex: bool,
@@ -21,10 +21,10 @@ pub fn broadcast_events(
         panic!("No relays specified, at least one relay is required!")
     }
 
-    let keys = handle_keys(None, sub_command_args.hex)?;
+    let keys = handle_keys(None, sub_command_args.hex, false)?;
     let client = create_client(&keys, relays.clone(), 0)?;
 
-    let file = std::fs::File::open(&sub_command_args.input)?;
+    let file = std::fs::File::open(&sub_command_args.file_path)?;
 
     let events: Vec<Event> = serde_json::from_reader(file)?;