+++ src/sub_commands/mute_publickey.rs
@@ -25,7 +25,7 @@ pub fn mute_publickey(
         panic!("No relays specified, at least one relay is required!")
     }
 
-    let keys = handle_keys(private_key, sub_command_args.hex)?;
+    let keys = handle_keys(private_key, sub_command_args.hex, true)?;
     let client = create_client(&keys, relays, difficulty_target)?;
 
     // Set up pubkey to mute