+++ src/sub_commands/send_channel_message.rs
@@ -25,7 +25,7 @@ pub fn send_channel_message(
         panic!("No relays specified, at least one relay is required!")
     }
 
-    let keys = handle_keys(private_key, sub_command_args.hex)?;
+    let keys = handle_keys(private_key, sub_command_args.hex, true)?;
     let client = create_client(&keys, relays.clone(), difficulty_target)?;
 
     let hex_channel_id: String = parse_key(sub_command_args.channel_id.clone())?;