+++ src/sub_commands/update_metadata.rs
@@ -36,7 +36,7 @@ pub fn update_metadata(
         panic!("No relays specified, at least one relay is required!")
     }
 
-    let keys = handle_keys(private_key, sub_command_args.hex)?;
+    let keys = handle_keys(private_key, sub_command_args.hex, true)?;
     let client = create_client(&keys, relays, difficulty_target)?;
 
     let mut metadata = Metadata::new();